from_mqtt/mqtt_cfs/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.1
